+++ zds/antispam/management/commands/antispam_train.py
@@ -1,0 +1,13 @@
+from django.core.management.base import BaseCommand
+
+from zds.antispam.spam_model_manager import SpamModelManager
+
+
+class Command(BaseCommand):
+    help = "Retrain the spam filter model and save it to a file."
+
+    def handle(self, *args, **options):
+        self.stdout.write("Starting retraining of the spam filter model...")
+        model_manager = SpamModelManager()
+        model_manager.retrain()
+        self.stdout.write("Retraining completed successfully.")
