+++ zds/settings/abstract_base/django.py
@@ -180,6 +180,7 @@ INSTALLED_APPS = (
     "zds.featured",
     "zds.search",
     "zds.notification",
+    "zds.antispam",
     # Uncomment the next line to enable the admin:
     "django.contrib.admin",
     # Uncomment the next line to enable admin documentation:
